fix(build): move ai prompt page to follow nextjs file convention
