--- app/ai_prompt/index.ts
@@ -1,3 +1,0 @@
-import { Page } from './page';
-
-export { Page as AIPromptPage };
